Refactor window logic
